Rebuild caches on changes to entries or navs
This refactors the caching into its own PageCache class for convenience.

This also means this cache can effectively live forever because it will be completely and correctly rebuilt whenever a entry or nav changes, which means only the CP editor who made the change will pay the penalty of building those caches, but only for those navs/entries.

It's the best of all worlds and depending on how your nav tags are configured, can shave hundreds of milliseconds of requests by visitors.

It will even pre-emptively build caches for pages that have just been created!